requirements.txt
streamlit
opencv-python-headless
pillow
pandas
numpy
easyocr